Ini Editor Done, Sprite Generator Done
Well the INI Editor works for what I need it to, I might expand it in
the future but for now it works fine. The sprite generator underwent a
massive overhaul, it went from a one purpose tool to now being able to
be used more than once. Oh and there's a crap ton of files because I
have yet to update the git ignore. I also started work on the A2XT
configuration for the NPC Editor