started work on some of the controls
yes, i'm doing each group of controls as one user control. this way
everything can be spaced evenly alot easier.

each group contains 3 items: the label, the value, and the checkbox. the
value can be either a combobox (wip) or a numeric spinner or even a
textbox. the checkbox signifies whether or not the value is enabled or
disabled for when you write, etc.