alot of work done
Well I changed a couple things, as of right now I've taken off the
animator class thing. I'm starting on working out the dynamic assigning
of the NpcConfig's key values based on events and stuff, it works pretty
well so far other than the fact that when incrementing the console can
get kind of flooded. OHHHH well, I'll probably find a way to filter that
out.

Most of the controls are done by now, I just have to set up the tagging
system for the Combo and CheckBox controls. I'm using the tags to
signify the the key that'll be assigned to, take a peek at the source to
see how its being used.